Rewrite hardware so simulation is closer to true hardware
We now push the register level control down to the simulation.  This now
means that the hardware simulation needs to be active during register
database loading.

In this commit the emulation of bit and position bus readback is now
broken, so the transcript test now fails on this part of the tests.