Rework top level Makefile
The main change is to remove the facility to build all applications at
once.  This can be restored with a special target or a helper script.

The makefile now manages the relationship between the stages of the build
more carefully, and for example `make zpkg` now builds only what needs to
be built.