Add missing parameters in IP definitions
In Zynq UltraScale+, some parameters are mandatory or have conflicting
defaults.
The default fifo implentation is common clock native, which does
not support having a data count output.
Another difference is that the 'asymetric port width' flag is needed when
the input data width and the output data width don't match.