Merge zpkgs into a single artifact to fix release
Production testing on the PandABlocks-server repo showed that the release action struggles to identify multiple zip files. The solution, which has become standard practice for a lot of diamond repos, is to upload to a single artifact that containins everything you wish to release. Then you can download this artifact in the release job and reference it the release action, which extracts the zip and uses the files in its release.