Fix wrong index stepping in dma interface
Each instance consumes a dma interface, therefore the dma index should be
stepped by the number of instances.

The bug was not manifested because there was only one module using the
dma interface, but soon, we will have more than one dma module and this
bug would be on our way.