Add a new interrupt for long tables
There are 2 cases in which a long table will interrupt: it became ready to
accept a new buffer or the block finished operation(and driver can release
resources). The interrupt is wired but not yet used by the blocks.

`dma_irq_events` was layed out conveniently for future latching in a
register.

Regarding the type of interrupt, usually, we shouldn't share edge
triggered interrupts, but in this case it's fine because:
1. The interrupt sources are synchronous to the same clock and therefore,
   the effect of many interrupts happening at the same time is possibly a
   wider pulse (with a very limited width).
2. By the time the interrupt routine is running, all interrupts will have
   been registered in the the status register(not yet implemented).

Device tree was updated in each target so that the next available SPI from
the GIC is used.