Improve and optimise streaming table implementation
The includes suggestions made in the last requirement meeting:
- MA: The DMA part (table_read_engine_client) shouldn't know about
repeats.
Therefore a new signal was introduced called `loop_one_buffer`.
- TC: Can we be as fast as before for small tables?
To implement this, instead of adding more block ram and multiplexing the
DMA output (which also requires a lot of complexity on the server side),
a more "integrated" approach was done, the fifo now has a
"wrapping_mode" that will allow to repeat the table (if it is < 4K
entries) without DMA intervention. This requires to have a custom
table/fifo entity which shouldn't be reused elsewhere, but it is worth
it.
- Tests were adapted accordingly.