Remove zpkg building and dependency on rootfs (#248)
* Remove zpkg building and dependency on rootfs

* Simplify app naming

Now application names contains only lower-case letters and - symbol.
This is done for 2 reasons:
- Reducing user typing mistakes.
- Matching the associated yocto package name (which has restrictive
naming rules).

* Remove fpga loading script as this is now part of the panda yocto layer

Additionally, target-defs has been moved to the target directory,
eliminating the need for the etc subfolder.

* Rename targets to match the same rules as app naming

This is for consistency and because in some context, the name of the
target is assumed to be the first part of the app name.

* autogen: add app name validation

* ipk: remove name sanitization which is not longer needed